scripts: Move atvremote
