+++ docs/api/pyatv/interface.html
@@ -1690,6 +1690,10 @@ def version(self) -&gt; Optional[str]:
 <ul class="hlist">
 <li>abc.ABC</li>
 </ul>
+<h3>Subclasses</h3>
+<ul class="hlist">
+<li>pyatv.scripts.atvremote.DeviceListener</li>
+</ul>
 <h3>Methods</h3>
 <dl>
 <dt id="pyatv.interface.DeviceListener.connection_closed"><code class="name flex">
@@ -2677,6 +2681,10 @@ def powerstate_update(
 <h3>Ancestors</h3>
 <ul class="hlist">
 <li>abc.ABC</li>
+</ul>
+<h3>Subclasses</h3>
+<ul class="hlist">
+<li>pyatv.scripts.atvremote.PushListener</li>
 </ul>
 <h3>Methods</h3>
 <dl>
